+++ lib/src/widget/asset_picker.dart
@@ -53,6 +53,7 @@ class AssetPicker<Asset, Path> extends StatefulWidget {
     PermissionRequestOption? permissionRequestOption,
     AssetPickerConfig pickerConfig = const AssetPickerConfig(),
     bool useRootNavigator = true,
+    RouteSettings? pageRouteSettings,
     AssetPickerPageRouteBuilder<List<AssetEntity>>? pageRouteBuilder,
   }) {
     return _pickerDelegate.pickAssets(
@@ -61,6 +62,7 @@ class AssetPicker<Asset, Path> extends StatefulWidget {
       pickerConfig: pickerConfig,
       permissionRequestOption: permissionRequestOption,
       useRootNavigator: useRootNavigator,
+      pageRouteSettings: pageRouteSettings,
       pageRouteBuilder: pageRouteBuilder,
     );
   }
@@ -73,6 +75,7 @@ class AssetPicker<Asset, Path> extends StatefulWidget {
     PermissionRequestOption permissionRequestOption =
         const PermissionRequestOption(),
     Key? key,
+    RouteSettings? pageRouteSettings,
     AssetPickerPageRouteBuilder<List<Asset>>? pageRouteBuilder,
     bool useRootNavigator = true,
   }) {
@@ -82,6 +85,7 @@ class AssetPicker<Asset, Path> extends StatefulWidget {
       delegate: delegate,
       permissionRequestOption: permissionRequestOption,
       useRootNavigator: useRootNavigator,
+      pageRouteSettings: pageRouteSettings,
       pageRouteBuilder: pageRouteBuilder,
     );
   }